Implement Paper API patches 0011-0020 - Add command line option to load extra plugin jars - Implement Player affects spawning API - Add getTPS method - Implement Version Command 2.0 - Add Entity Origin API - Implement PlayerLocaleChangeEvent - Add View distance API - Implement BeaconEffectEvent - Add PlayerInitialSpawnEvent - Expose server CommandMap

+++ arclight-common/src/main/java/io/izzel/arclight/common/mixin/core/world/entity/ai/behavior/BabyFollowAdultMixin.java
@@ -34,7 +34,7 @@ public abstract class BabyFollowAdultMixin {
                         return false;
                     } else {
                         LivingEntity ageablemob = p_258331_.get(p_258317_);
-                        if (p_258327_.closerThan(ageablemob, (double) (p_259321_.getMaxValue() + 1)) && !p_258327_.closerThan(ageablemob, (double) p_259321_.getMinValue())) {
+                        if (p_258327_.closerThan(ageablemob, p_259321_.getMaxValue() + 1) && !p_258327_.closerThan(ageablemob, p_259321_.getMinValue())) {
                             // CraftBukkit start
                             EntityTargetLivingEntityEvent event = CraftEventFactory.callEntityTargetLivingEvent(p_258327_, ageablemob, EntityTargetEvent.TargetReason.FOLLOW_LEADER);
                             if (event.isCancelled()) {
